requirements.txt
--index-url https://pypi.org/simple

--extra-index-url https://download.pytorch.org/whl/cpu

pandas
numpy<2.0
scikit-learn
deepface
tensorflow-cpu
tf-keras


torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu


sentence-transformers
faiss-cpu
catboost
Pillow
tqdm
sqlalchemy